pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gmsatool"
description = "A gMSA tool for enumeration, access management, and password retrieval."
version = "0.1.0"
requires-python = ">= 3.11"
dependencies = [
    "gssapi>=1.10.1",
    "ldap3-bleeding-edge>=2.10.1.1338",
    "pycryptodome>=3.23.0",
    "typer>=0.20.0",
    "winacl>=0.1.9",
]
authors = [
  {name = "Niket", email = "niket5151@gmail.com"},
]
readme = "README.md"
license="MIT"
classifiers=[
  'Programming Language :: Python :: 3',
]

[project.urls]
homepage = "https://github.com/int3x/gmsatool"
repository = "https://github.com/int3x/gmsatool"

[project.scripts]
gmsatool = "main:app"

[tool.ruff]
line-length = 200
